requirements.txt
requests>=2.22.0
tqdm>=4.36.1
beautifulsoup4>=4.8.0
PyQt5>=5.14.1
scipy>=1.3.1
tensorflow>=2.0.0
numpy>=1.17.2
Pillow>=6.2.0
torch>=1.3.1
torchvision>=0.4.2
